attempt to fix

+++ frontend/lib/Pages/service/database.dart
@@ -1,3 +1,5 @@
+import 'dart:ffi';
+
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -21,7 +23,7 @@ class UserModel {
   int avatar;
   int character;
   String email;
-  int hp;
+  double  hp;
   int level;
   String name;
   int nutrition;
